Rename stats modal to panel
