+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/CustomBarChart.tsx
