+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/CustomChartTab.tsx
