+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/MatrixAxisSelector.tsx
