+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/MatrixFilters.tsx
