+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/MatrixSection.tsx
