+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/NoDataPlaceholder.tsx
