+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/StatsCard.tsx
