+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/constants.tsx
@@ -5,7 +5,7 @@ import QueryStatsIcon from '@mui/icons-material/QueryStats';
 import InsightsIcon from '@mui/icons-material/Insights';
 import BarChartIcon from '@mui/icons-material/BarChart';
 import DashboardCustomizeIcon from '@mui/icons-material/DashboardCustomize';
-import type { StatsView } from '../StatsModal';
+import type { StatsView } from '../StatsPanel';
 
 export const TAB_DEFINITIONS: ReadonlyArray<{
   value: StatsView;
