+++ src/features/videoPlayer/components/Analytics/StatsPanel/view/utils.ts
